Alanna's edit to Mix's education article
Wow. This is a gorgeous, powerful piece of writing. It packs both an intellectual and personal punch. I've made some changes to make it a bit tighter in some areas, and expanded a bit in others to make it more accessible. Keep going - this is very worth polishing and publishing.